new config post merge fixes - move ai consts
