+++ extensions/void/src/ai/constants.ts
@@ -1,0 +1,11 @@
+export enum ApiProvider {
+	ANTHROPIC = 'anthropic',
+	OPENAI = 'openai',
+	GREPTILE = 'greptile',
+	OLLAMA = 'ollama',
+	OPENAI_COMPATIBLE = 'openAICompatible'
+}
+
+export enum VectorStore {
+	OPENSEARCH = 'opensearch'
+}
